# panel/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "3X-UI Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a 3X-UI panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "3x-ui",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/panel/api/inbounds/add": {
      "post": {
        "tags": [
          "Entrypoints"
        ],
        "summary": "Open a new entrypoint on a node. The port must be free on that node — the check is per node, not global, since two nodes may both listen on 443. Draft edit.",
        "operationId": "post_panel_api_inbounds_add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "nodeId": 1,
                "port": 8443,
                "remark": "spare 8443",
                "enable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "msg": "Port 8443 is already used on this node"
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/clients/list/paged": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Filter, sort and paginate clients server-side. Search matches the tag and the comment only. Sortable columns: createdAt, updatedAt, lastOnline, email (tag), enable, group, traffic, expiryTime. Rows are slim — fetch /get/:email for the full record. The summary counters are computed over the whole table, not the page, so they hold still while paging.",
        "operationId": "get_panel_api_clients_list_paged",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "description": "1-indexed page number. Defaults to 1.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "description": "Rows per page. Defaults to 25, capped at 200.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": true,
            "description": "Case-insensitive substring match on the tag and the comment. Keys are never searched.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "description": "Status bucket: online | active | deactive | expiring.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": true,
            "description": "createdAt | updatedAt | lastOnline | email | enable | group | traffic | expiryTime.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": true,
            "description": "ascend or descend.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "items": [
                      {
                        "email": "alice@example.com",
                        "subId": "abcd1234",
                        "enable": true,
                        "totalGB": 53687091200,
                        "expiryTime": 1735689600000,
                        "limitIp": 0,
                        "reset": 0,
                        "inboundIds": [
                          3,
                          5
                        ],
                        "traffic": {
                          "up": 1024,
                          "down": 4096,
                          "enable": true
                        },
                        "createdAt": 1735000000000,
                        "updatedAt": 1735100000000
                      }
                    ],
                    "total": 2000,
                    "filtered": 47,
                    "page": 1,
                    "pageSize": 25,
                    "summary": {
                      "total": 2000,
                      "active": 1850,
                      "online": [
                        "alice@example.com"
                      ],
                      "depleted": [],
                      "expiring": [],
                      "deactive": []
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/clients/add": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Create a client. The key (uuid) is generated when omitted. This writes to the draft — the client cannot connect anywhere until the draft is published to the nodes.",
        "operationId": "post_panel_api_clients_add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "client": {
                  "email": "alice@example.com",
                  "totalGB": 53687091200,
                  "expiryTime": 1735689600000,
                  "tgId": 0,
                  "limitIp": 0,
                  "enable": true
                },
                "inboundIds": [
                  3,
                  5
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "msg": "Client added"
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/clients/groups/create": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Create an empty group. It becomes selectable immediately, but a group without entrypoints grants nothing — add entrypoints with /groups/entrypoints.",
        "operationId": "post_panel_api_clients_groups_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "customer-a"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "name": "customer-a"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/clients/groups/delete": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Remove a group and clear it from every client that carried it. The clients themselves are kept — filter by group and use /bulkDel to remove them. Returns how many were cleared.",
        "operationId": "post_panel_api_clients_groups_delete",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "customer-a"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "affected": 5
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/nodes/add": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Record a new node and mint the admin key only this node will accept. Nothing is contacted — the machine does not exist yet. The Add dialog turns the same values into a deploy script; running it on the new machine is what brings the node up, after which it reports in and leaves the \"waiting\" state.",
        "operationId": "post_panel_api_nodes_add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "node-9",
                "role": "egress",
                "address": "node-9.example.com",
                "port": 443,
                "apiToken": "7f3a…",
                "enable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}